arrange the project structure for the preparation of crates publish; users can call crates more Conveniently
