+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
 name = "curs"
-version = "0.3.1"
+version = "0.3.2"
 
 [dependencies]
 indicatif = "*"
