+++ src/query/mod.rs
@@ -1,0 +1,11 @@
+mod cli;
+mod extractor;
+mod extractor_chooser;
+mod files;
+mod language;
+
+pub use cli::{Invocation, QueryFormat, QueryOpts};
+pub use extractor::{ExtractedFile, ExtractedMatch, Extractor};
+pub use extractor_chooser::ExtractorChooser;
+pub use files::Files;
+pub use language::Language;
