+++ src/safe/mod.rs
@@ -1,0 +1,3 @@
+mod inference;
+
+pub use inference::{classify, do_query, show_languages};
